+++ test/attr-interpolation.js
@@ -1,4 +1,3 @@
-var Emitter = require('emitter');
 var domify = require('domify');
 var assert = require('assert');
 
@@ -22,27 +21,23 @@ describe('attr interpolation', function(){
   it('should react to changes', function(){
     var el = domify('<a href="/download/{id}"></a>');
     var user = { id: '1234' };
-    Emitter(user);
     var view = reactive(el, user);
 
     assert('/download/1234' == el.getAttribute('href'));
 
-    user.id = '4321';
-    user.emit('change id');
+    view.set('id', '4321');
     assert('/download/4321' == el.getAttribute('href'));
   })
 
   it('should support multiple attributes', function(){
     var el = domify('<a href="/download/{id}" id="file-{id}">Download {file}</a>');
     var user = { id: '1234' };
-    Emitter(user);
     var view = reactive(el, user);
 
     assert('/download/1234' == el.getAttribute('href'));
     assert('file-1234' == el.getAttribute('id'));
 
-    user.id = '4321';
-    user.emit('change id');
+    view.set('id', '4321');
     assert('/download/4321' == el.getAttribute('href'));
     assert('file-4321' == el.getAttribute('id'));
   })
@@ -50,17 +45,14 @@ describe('attr interpolation', function(){
   it('should support multiple properties', function(){
     var el = domify('<a href="/download/{id}-{file}"></a>');
     var user = { id: '1234', file: 'something' };
-    Emitter(user);
     var view = reactive(el, user);
 
     assert('/download/1234-something' == el.getAttribute('href'));
 
-    user.id = '4321';
-    user.emit('change id');
+    view.set('id', '4321');
     assert('/download/4321-something' == el.getAttribute('href'));
 
-    user.file = 'whoop';
-    user.emit('change file');
+    view.set('file', 'whoop');
     assert('/download/4321-whoop' == el.getAttribute('href'));
   })
 })