+++ History.md
@@ -5,6 +5,7 @@ unreleased
  * remove explicit model method calls and view method calls in interpolation bindings
  * remove global `bindings` and `use`. Set on instances instead
  * add `Reactive#destroy` method to facilitate destroying the view and unbinding handlers
+ * add `each` binding for iteration support
 
 0.14.1 / 2014-01-09
 ===================