+++ History.md
@@ -1,4 +1,4 @@
-unreleased
+1.0.0 / 2014-03-05
 ==========
  * contain adapter within an Adapter class to allow per-instance adapters
  * remove change event binding from Adapter (leave to custom adapters)